requirements.txt
flask
flask-sqlalchemy
flask-cors
werkzeug
google-generativeai
opencv-python-headless
deepface
numpy
scikit-learn
spacy
nltk
textblob
SpeechRecognition
librosa
soundfile
PyPDF2
python-docx
pandas
matplotlib
reportlab
Pillow
python-dotenv
requests
gunicorn
